+++ src/mass/spec/SavitzkyGolay.java
@@ -17,43 +17,43 @@ import org.apache.commons.math3.optim.nonlinear.vector.jacobian.LevenbergMarquar
  */
 
 public class SavitzkyGolay {
-    double[][] data;
+    float[][] data;
     int degree;
     int window;
     
-        public SavitzkyGolay (double[][] vals, int wind, int deg) {
+        public SavitzkyGolay (float[][] vals, int wind, int deg) {
             if (window%2 != 1) window-=1;
             window = wind;
             degree = deg;
-            data = new double[2][];
+            data = new float[2][];
             data[0] = Arrays.copyOf(vals[0], vals[0].length);
             data[1] = Arrays.copyOf(vals[1], vals[1].length);
         }
         
-        public SavitzkyGolay (double[] xvals, double[] yvals, int wind, int deg) {
+        public SavitzkyGolay (float[] xvals, float[] yvals, int wind, int deg) {
             if (window%2 != 1) window-=1;
             window = wind;
             degree = deg;
-            data = new double[2][];
+            data = new float[2][];
             data[0] = Arrays.copyOf(xvals, xvals.length);
             data[1] = Arrays.copyOf(yvals, yvals.length);
         }
         
-        public double getSmoothedPoint (int index) {
+        public float getSmoothedPoint (int index) {
             if (index < window) index = window;
             if (index > data[0].length - window) index = data[0].length - window;
-            double[][] values = new double [2][];
+            float[][] values = new float [2][];
             values[0] = Arrays.copyOfRange(data[0], index - window, index + window);
             values[1] = Arrays.copyOfRange(data[1], index - window, index + window);
             double[] parameters = fitPolynomial(values);
-            double sum = 0;
+            float sum = 0;
             for (int i = 0; i < degree; i++) {
                 sum += parameters[i]*Math.pow(data[0][index], i);
             }
             return sum;
         }
         
-        private double[] fitPolynomial (double[][] data) {
+        private double[] fitPolynomial (float[][] data) {
             PolynomialFitter fitter = new PolynomialFitter(new LevenbergMarquardtOptimizer());
              for (int i = 0; i < data[0].length; i++) {
                  fitter.addObservedPoint(data[0][i], data[1][i]);
